pyproject.toml
[project]
name = "fabric-cicd"
authors = [{ name = "Microsoft Corporation" }]
description = "Microsoft Fabric CI/CD"
readme = "README.md"
requires-python = ">=3.9,<3.13"
license = { file = "LICENSE" }
dynamic = ["version"]
urls.Repository = "https://github.com/microsoft/fabric-cicd.git"

dependencies = [
    "azure-identity>=1.19.0",
    "dpath>=2.2.0",
    "filetype>=1.2.0",
    "ms-fabric-cli>=0.1.10 ; python_full_version >= '3.10'",
    "packaging>=24.2",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "coverage>=7.6.10",
    "gitpython>=3.1.44",
    "mike>=2.1.3",
    "mkdocs-include-markdown-plugin>=7.1.2",
    "mkdocs-material>=9.6.5",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocstrings-python>=1.13.0",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
dynamic.version = { attr = "fabric_cicd.constants.VERSION" }
packages.find.where = ["src"]
package-data = { "fabric_cicd" = ["changelog.md"] }

[tool.uv]
package = true
python-preference = "only-managed"

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
